More Direct3D 10 samples

+++ Direct3D10/AdvancedParticles/AdvancedParticles.cpp
@@ -340,9 +340,9 @@ void CALLBACK OnFrameMove( double fTime, float fElapsedTime, void* pUserContext 
     {
         if( fTime - fLastRandomize > 2.0 )
         {
-            g_vParticleColor.x = fabs( RPercent() );
-            g_vParticleColor.y = fabs( RPercent() );
-            g_vParticleColor.z = fabs( RPercent() );
+            g_vParticleColor.x = fabsf( RPercent() );
+            g_vParticleColor.y = fabsf( RPercent() );
+            g_vParticleColor.z = fabsf( RPercent() );
 
             fLastRandomize = fTime;
         }
